% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_metadata.r
\name{output_meta_data_fx}
\alias{output_meta_data_fx}
\title{create txt file with metadata for r output written to disk}
\usage{
output_meta_data_fx(
  output_file_nm,
  script.dir = script.dir.info,
  script.name = script.name.info,
  get.script.line.number = T,
  notes = ""
)
}
\arguments{
\item{output_file_nm}{directory and file name of output}

\item{script.dir}{directory of code repository (place it on top of r script)}

\item{notes}{any text string to be added to the text file}

\item{script.name.info}{name of script used to generate it (place it on top of r script)}
}
\value{
text file with same name as output file with a lot of metadata info
}
\description{
This function creates a txt file with metadata for any r output written to disk.
The output should help reproducing the analysis as it lists r script file info
along with info workstation info.
}
